requirements.txt
selenium
python-dotenv
duckdb